+++ packages/react-data-grid-react-window/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-contrib/react-data-grid-react-window
 
-This log was last generated on Tue, 03 Dec 2024 12:55:55 GMT and should not be manually modified.
+This log was last generated on Tue, 17 Dec 2024 11:27:26 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.3.1
+
+Tue, 17 Dec 2024 11:27:26 GMT
+
+### Patches
+
+- refactor (react-data-grid-react-window): export all types and hooks (alton@microsoft.com)
 
 ## 1.3.0